git-status.sh: First commit date WIP
Work in progress.

Working on showing a graph of the number of days since the first commit
for each committer. Getting hung up on the commit date for some reason.
It looks like it's exiting after getting the first date and I'm not sure
why. Going to commit what I have and go to bed or something and look at
this some other time.